refactor(neurome): add NeuRome configuration
Now models should be able to run either on ROOT-Sim 2 (configured by the rootsim-setup script) or by
ROOT-Sim 3 a.k.a. NeuRome (configured by the neurome-setup script). Binaries of either configuration
(only one at a time) are found in the Simulator folder, models have already made been compliant with
the new binary folder

BREAKING CHANGE: The ROOT-Sim-bin folder eos not exist anymore, the new location of rootsim-cc is in
Simulator/bin and the new location of ROOT-Sim.h is Simulator/include

Signed-off-by: Nick1296 <mattianicolella@gmail.com>