feat(new_pipeline): complete rework of the pipeline to execute simulations
With the new pipeline each supported simulator is downloaded and configured inside "Simulators".
Then each component of the pipeline will need some additional arguments to locate the input files
and to decide where the output file must be located. Compilation of the tree_simulator model now
only uses the makefile (run.sh is now a simple wrapper to avoid manual argument insertion); location
of the executables can be changed with the LOCATION variable and debug options (-g -O0) can be
triggere with the ENABLE_DBG_OPTS variable (both are optional and must be passed before the target).
Finally the pipeline can be borken down in segments, on for each step (two when we are compiling and
executing the model).

BREAKING CHANGE: All the scripts take parameters to customize their output files location.
application_datatypes.h now has only datatypes for the simulation model, topology datatypes are in
partop_datatypes and scheduler datatypes are in scheduler_datatypes.

Signed-off-by: Nick1296 <mattianicolella@gmail.com>